pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tooledit"
version = "0.1.0"
description = "Auditable non-generative image editing orchestration with LangGraph."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" },
]
dependencies = [
  "google-genai>=1.11.0",
  "langgraph>=0.2.50",
  "numpy>=1.26.0",
  "opencv-python-headless>=4.10.0.84",
  "pillow>=10.4.0",
  "pydantic>=2.9.0",
  "rich>=13.9.0",
  "scikit-image>=0.24.0",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
tooledit = "tooledit.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/tooledit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
